Pequeñas correcciones y añadidos.

+++ source/calculos1Var.hpp
@@ -1,7 +1,7 @@
-/***************************************************************************														*
- *   Copyright (C) 2018 by Ángel Bravo Sáenz  										*
- *   angelbravosaenz@gmail.com  															*
- *																									*
+/***************************************************************************
+ *   Copyright (C) 2018 by Ángel Bravo Sáenz                               *
+ *   angelbravosaenz@gmail.com                                             *
+ *                                                                         *
  *   This file is part of Stac.                                            *
  *                                                                         *
  *   Stac is free software; you can redistribute it and/or modify          *
@@ -31,7 +31,7 @@ class Calculos1Var : public Vector {
 		//======== constructor ======== 
 		Calculos1Var (bool leerArchivo, unsigned int longitudVector);
 		//========  funciones de estadística descriptiva ======== 
-		double sumatoria ( short unsigned int exponente );
+		double sumatoria ( unsigned int exponente );
 		double mediaAritmetica();
 		double mediaGeometrica() ;
 		double mediaArmonica();
